restic: extract Node filesystem code to fs package
