+++ internal/fs/node_aix.go
@@ -1,11 +1,13 @@
 //go:build aix
 // +build aix
 
-package restic
+package fs
 
 import (
 	"os"
 	"syscall"
+
+	"github.com/restic/restic/internal/restic"
 )
 
 func nodeRestoreSymlinkTimestamps(_ string, _ [2]syscall.Timespec) error {
@@ -24,12 +26,12 @@ func (s statT) mtim() syscall.Timespec { return toTimespec(s.Mtim) }
 func (s statT) ctim() syscall.Timespec { return toTimespec(s.Ctim) }
 
 // nodeRestoreExtendedAttributes is a no-op on AIX.
-func nodeRestoreExtendedAttributes(_ *Node, _ string) error {
+func nodeRestoreExtendedAttributes(_ *restic.Node, _ string) error {
 	return nil
 }
 
 // nodeFillExtendedAttributes is a no-op on AIX.
-func nodeFillExtendedAttributes(_ *Node, _ string, _ bool) error {
+func nodeFillExtendedAttributes(_ *restic.Node, _ string, _ bool) error {
 	return nil
 }
 
@@ -39,11 +41,11 @@ func IsListxattrPermissionError(_ error) bool {
 }
 
 // nodeRestoreGenericAttributes is no-op on AIX.
-func nodeRestoreGenericAttributes(node *Node, _ string, warn func(msg string)) error {
-	return HandleAllUnknownGenericAttributesFound(node.GenericAttributes, warn)
+func nodeRestoreGenericAttributes(node *restic.Node, _ string, warn func(msg string)) error {
+	return restic.HandleAllUnknownGenericAttributesFound(node.GenericAttributes, warn)
 }
 
 // nodeFillGenericAttributes is a no-op on AIX.
-func nodeFillGenericAttributes(_ *Node, _ string, _ os.FileInfo, _ *statT) (allowExtended bool, err error) {
+func nodeFillGenericAttributes(_ *restic.Node, _ string, _ os.FileInfo, _ *statT) (allowExtended bool, err error) {
 	return true, nil
 }
