+++ internal/fs/node_netbsd.go
@@ -1,8 +1,10 @@
-package restic
+package fs
 
 import (
 	"os"
 	"syscall"
+
+	"github.com/restic/restic/internal/restic"
 )
 
 func nodeRestoreSymlinkTimestamps(_ string, _ [2]syscall.Timespec) error {
@@ -14,12 +16,12 @@ func (s statT) mtim() syscall.Timespec { return s.Mtimespec }
 func (s statT) ctim() syscall.Timespec { return s.Ctimespec }
 
 // nodeRestoreExtendedAttributes is a no-op on netbsd.
-func nodeRestoreExtendedAttributes(_ *Node, _ string) error {
+func nodeRestoreExtendedAttributes(_ *restic.Node, _ string) error {
 	return nil
 }
 
 // nodeFillExtendedAttributes is a no-op on netbsd.
-func nodeFillExtendedAttributes(_ *Node, _ string, _ bool) error {
+func nodeFillExtendedAttributes(_ *restic.Node, _ string, _ bool) error {
 	return nil
 }
 
@@ -29,11 +31,11 @@ func IsListxattrPermissionError(_ error) bool {
 }
 
 // nodeRestoreGenericAttributes is no-op on netbsd.
-func nodeRestoreGenericAttributes(node *Node, _ string, warn func(msg string)) error {
-	return HandleAllUnknownGenericAttributesFound(node.GenericAttributes, warn)
+func nodeRestoreGenericAttributes(node *restic.Node, _ string, warn func(msg string)) error {
+	return restic.HandleAllUnknownGenericAttributesFound(node.GenericAttributes, warn)
 }
 
 // nodeFillGenericAttributes is a no-op on netbsd.
-func nodeFillGenericAttributes(_ *Node, _ string, _ os.FileInfo, _ *statT) (allowExtended bool, err error) {
+func nodeFillGenericAttributes(_ *restic.Node, _ string, _ os.FileInfo, _ *statT) (allowExtended bool, err error) {
 	return true, nil
 }
