+++ internal/fs/node_xattr.go
@@ -1,7 +1,7 @@
 //go:build darwin || freebsd || linux || solaris
 // +build darwin freebsd linux solaris
 
-package restic
+package fs
 
 import (
 	"fmt"
@@ -10,6 +10,7 @@ import (
 
 	"github.com/restic/restic/internal/debug"
 	"github.com/restic/restic/internal/errors"
+	"github.com/restic/restic/internal/restic"
 
 	"github.com/pkg/xattr"
 )
@@ -65,16 +66,16 @@ func handleXattrErr(err error) error {
 }
 
 // nodeRestoreGenericAttributes is no-op.
-func nodeRestoreGenericAttributes(node *Node, _ string, warn func(msg string)) error {
-	return HandleAllUnknownGenericAttributesFound(node.GenericAttributes, warn)
+func nodeRestoreGenericAttributes(node *restic.Node, _ string, warn func(msg string)) error {
+	return restic.HandleAllUnknownGenericAttributesFound(node.GenericAttributes, warn)
 }
 
 // nodeFillGenericAttributes is a no-op.
-func nodeFillGenericAttributes(_ *Node, _ string, _ os.FileInfo, _ *statT) (allowExtended bool, err error) {
+func nodeFillGenericAttributes(_ *restic.Node, _ string, _ os.FileInfo, _ *statT) (allowExtended bool, err error) {
 	return true, nil
 }
 
-func nodeRestoreExtendedAttributes(node *Node, path string) error {
+func nodeRestoreExtendedAttributes(node *restic.Node, path string) error {
 	expectedAttrs := map[string]struct{}{}
 	for _, attr := range node.ExtendedAttributes {
 		err := setxattr(path, attr.Name, attr.Value)
@@ -101,7 +102,7 @@ func nodeRestoreExtendedAttributes(node *Node, path string) error {
 	return nil
 }
 
-func nodeFillExtendedAttributes(node *Node, path string, ignoreListError bool) error {
+func nodeFillExtendedAttributes(node *restic.Node, path string, ignoreListError bool) error {
 	xattrs, err := listxattr(path)
 	debug.Log("fillExtendedAttributes(%v) %v %v", path, xattrs, err)
 	if err != nil {
@@ -111,14 +112,14 @@ func nodeFillExtendedAttributes(node *Node, path string, ignoreListError bool) e
 		return err
 	}
 
-	node.ExtendedAttributes = make([]ExtendedAttribute, 0, len(xattrs))
+	node.ExtendedAttributes = make([]restic.ExtendedAttribute, 0, len(xattrs))
 	for _, attr := range xattrs {
 		attrVal, err := getxattr(path, attr)
 		if err != nil {
 			fmt.Fprintf(os.Stderr, "can not obtain extended attribute %v for %v:\n", attr, path)
 			continue
 		}
-		attr := ExtendedAttribute{
+		attr := restic.ExtendedAttribute{
 			Name:  attr,
 			Value: attrVal,
 		}
